// apps/foreldrepengeoversikt/src/app/types/TidslinjehendelseType.ts
export enum TidslinjehendelseType {
    FØRSTEGANGSSØKNAD = 'FØRSTEGANGSSØKNAD',
    FØRSTEGANGSSØKNAD_NY = 'FØRSTEGANGSSØKNAD_NY',
    ETTERSENDING = 'ETTERSENDING',
    ENDRINGSSØKNAD = 'ENDRINGSSØKNAD',
    INNTEKTSMELDING = 'INNTEKTSMELDING',
    VEDTAK = 'VEDTAK',
    VEDTAK_FØRSTEGANG = 'VEDTAK_FØRSTEGANG',
    VEDTAK_ENDRING = 'VEDTAK_ENDRING',
    VEDTAK_TILBAKEKREVING = 'VEDTAK_TILBAKEKREVING',
    VENTER_INNTEKTSMELDING = 'VENTER_INNTEKTSMELDING',
    VENTER_PGA_TIDLIG_SØKNAD = 'VENTER_PGA_TIDLIG_SØKNAD',
    VENTER_MELDEKORT = 'VENTER_MELDEKORT',
    VENT_DOKUMENTASJON = 'VENT_DOKUMENTASJON',
    UTGÅENDE_INNHENT_OPPLYSNINGER = 'UTGÅENDE_INNHENT_OPPLYSNINGER',
    UTGÅENDE_ETTERLYS_INNTEKTSMELDING = 'UTGÅENDE_ETTERLYS_INNTEKTSMELDING',
    FAMILIEHENDELSE = 'FAMILIEHENDELSE',
}